+++ src/gospatial/app/layerHandlers.go
@@ -3,6 +3,7 @@ package app
 import (
 	"encoding/json"
 	"github.com/gorilla/mux"
+	"github.com/paulmach/go.geojson"
 	"net/http"
 )
 
@@ -78,10 +79,12 @@ func NewLayerHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Create datasource
-	geojs := NewGeojson()
+	// geojs := NewGeojson()
 	ds, _ := NewUUID()
-	lyr := Layer{Datasource: ds, Geojson: geojs}
-	lyr.Save()
+	// lyr := Layer{Datasource: ds, Geojson: geojs}
+	// lyr.Save()
+	featCollection := geojson.NewFeatureCollection()
+	DB.insertLayer(ds, featCollection)
 
 	// Add datasource uuid to customer
 	customer.Datasources = append(customer.Datasources, ds)
@@ -157,7 +160,8 @@ func ViewLayerHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Marshal datasource layer to json
-	js, err := json.Marshal(lyr)
+	// js, err := json.Marshal(lyr)
+	rawJSON, err := lyr.MarshalJSON()
 	if err != nil {
 		Error.Println(r.RemoteAddr, "GET /api/v1/layer/"+ds+" [500]")
 		http.Error(w, err.Error(), http.StatusInternalServerError)
@@ -169,7 +173,7 @@ func ViewLayerHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	// allow cross domain AJAX requests
 	w.Header().Set("Access-Control-Allow-Origin", "*")
-	w.Write(js)
+	w.Write(rawJSON)
 
 }